Moved StoreListInput to dispatch/components
